Revert "ABM Habitaciones"
This reverts commit 4d53ca66bbdaef54e0440612e3de31e2006a5b78.

+++ Proyecto 1er año/EntidadesCompartidas/Habitacion.cs
@@ -125,7 +125,7 @@ namespace EntidadesCompartidas
         }
 
        //Constructor
-        public Habitacion(int pNumero, Hotel pHotel, string pPiso, string pDescripcion, string pHuespedes, double pCostodiario)
+        public Habitacion(int pNumero, Hotel pHotel, string pPiso, string pDescripcion, string pHuespedes, float pCostodiario)
         {
             Numero = pNumero;
             Hotel = pHotel;
